+++ tests/test_app/Plugin/TestPlugin/config/routes.php
@@ -1,0 +1,12 @@
+<?php
+declare(strict_types=1);
+
+use Cake\Routing\Route\DashedRoute;
+use Cake\Routing\RouteBuilder;
+
+/** @var \Cake\Routing\RouteBuilder $routes */
+$routes->setRouteClass(DashedRoute::class);
+
+$routes->scope('/', ['plugin' => 'TestPlugin'], function (RouteBuilder $routes) {
+    $routes->fallbacks();
+});