+++ tests/test_app/Plugin/TestPlugin/src/Controller/PagesController.php
@@ -18,4 +18,8 @@ use Cake\Controller\Controller;
 
 class PagesController extends Controller
 {
+    public function index()
+    {
+
+    }
 }